Decoupling of the Pagination component
